+++ NEWS.md
@@ -9,7 +9,9 @@ The current version of the `MSEtool` package is available for download from [CRA
 - F_crash, F_med, and other reference points have been added to `Hist@Ref$ByYear`
 - can now optimize for depletion in terms of SB/SBMSY with `control$D <- 'SBMSY'`
 - option to simulate CAL data including removals (retained + discards) with `OM@cpars$control$CAL="removals"`. 
-Default is to simulate only retained catch-at-length (default).
+Default is to simulate only retained catch-at-length.
+- option to apply TAC to all removals with `OM@cpars$control$TAC="removals"`. 
+Default is to for TAC to refer to retained catch (i.e., removals will be greater than TAC if there are discards)
 
 ### Minor changes
 - add more informative messages when conditioning on real data